backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
ruff>=0.4.0
